moves dev team data file and adds test
